indicator group refactor store
