+++ src/Module/DataElementGroup/Components/Facts/index.js
@@ -4,7 +4,7 @@ import Loader from "../../../../Shared/Componets/Loaders/Loader";
 import Error from "../../../../Shared/Componets/Error/ErrorAPIResult";
 import {dataSetDataElementCountState, programDataElementCountState} from "../../../../Store";
 import {useRecoilValue} from "recoil";
-import IndicatorCount from "./Components/IndicatorCount";
+import IndicatorCount from "../../../../Shared/Componets/IndicatorCount";
 
 const query = {
     sources:{
@@ -41,7 +41,7 @@ export default function Facts({id}){
         <ul>
             <li> It has {data?.sources?.dataElements?.length} data Elements     </li>
             <li>It’s data elements belongs to {dataSetCount} dataset and {programCount} program sources of data</li>
-            <li> {}
+            <li>
                 <IndicatorCount dataElements={data?.sources?.dataElements}/> </li>
         </ul>
     </div>
